Change to rightDigit value assignment conditions
Removing if(*rightDigitPtr <= *leftDigitPtr) resets rightDigit anytime
leftDigit is incremented this function, which solves a glitch we had
where rightDigit would increasingly raise in value, invalidating swap
comparions, and leading to an out-of-bounds error.

Placing the other rightDigit assignment within an else clause was simply
to save some processing cycles.